Fix log.Sink and add NoopSink (#172)
In the first version of the log.Sink implementation, a single log.Sink instance was registered on all incoming requests which leads to an effect that every sentry message contained all log messages. This pr fixes this behaviour and also introduces a log.NoopSink to be able to store logs in this sink while discarding them before reaching the logger.